api-gateway/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
requests>=2.31.0
vantage6-client>=4.13.0
